backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
torch
torchvision
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
faiss-cpu
pillow
pytesseract
opencv-python-headless
numpy
scikit-learn
insightface
onnxruntime
exifread
vosk
pyaudio
